UP-Fall/requirements.txt
absl-py @ file:///croot/absl-py_1666362940888/work
aiohttp @ file:///tmp/build/80754af9/aiohttp_1646788240124/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
astor==0.8.1
astunparse==1.6.3
async-timeout @ file:///opt/conda/conda-bld/async-timeout_1664876359750/work
asynctest==0.13.0
attrs @ file:///opt/conda/conda-bld/attrs_1642510447205/work
blinker==1.4
brotlipy==0.7.0
cachetools==5.2.0
certifi @ file:///croot/certifi_1665076670883/work/certifi
cffi @ file:///tmp/abs_98z5h56wf8/croots/recipe/cffi_1659598650955/work
charset-normalizer==2.1.1
click @ file:///tmp/build/80754af9/click_1646056591685/work
cmake==3.18.0
cryptography @ file:///croot/cryptography_1665612644927/work
Cython==0.29.32
flatbuffers==22.10.26
frozenlist @ file:///tmp/build/80754af9/frozenlist_1637767116045/work
gast @ file:///Users/ktietz/demo/mc3/conda-bld/gast_1628588903283/work
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta @ file:///Users/ktietz/demo/mc3/conda-bld/google-pasta_1630577991354/work
grpcio==1.50.0
h5py==3.7.0
idna @ file:///croot/idna_1666125576474/work
importlib-metadata==5.0.0
jax==0.3.24
keras==2.10.0
keras-nightly==2.12.0.dev2022111408
Keras-Preprocessing @ file:///tmp/build/80754af9/keras-preprocessing_1612283640596/work
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
multidict @ file:///croot/multidict_1665674239670/work
numpy==1.21.6
oauthlib==3.2.2
opencv-python==4.6.0.66
opt-einsum @ file:///tmp/build/80754af9/opt_einsum_1621500238896/work
packaging==21.3
Pillow==9.3.0
protobuf==3.19.6
pyasn1 @ file:///Users/ktietz/demo/mc3/conda-bld/pyasn1_1629708007385/work
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
PyJWT @ file:///opt/conda/conda-bld/pyjwt_1657544592787/work
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
PySocks @ file:///tmp/build/80754af9/pysocks_1594394576006/work
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
requests-oauthlib==1.3.1
rsa==4.9
scipy @ file:///opt/conda/conda-bld/scipy_1661390393401/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
tb-nightly==2.11.0a20221109
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit @ file:///home/builder/tkoch/workspace/tensorflow/tensorboard-plugin-wit_1658918494740/work/tensorboard_plugin_wit-1.8.1-py3-none-any.whl
tensorflow @ file:///home/builder/ktietz/aggregate/tensorflow_recipes/ci_cpu/tensorflow-base_1614534878484/work/tensorflow_pkg/tensorflow-2.4.1-cp37-cp37m-linux_x86_64.whl
tensorflow-estimator==2.10.0
tensorflow-gpu==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor @ file:///croot/termcolor_1668084651543/work
tf-estimator-nightly==2.12.0.dev2022111409
typing_extensions==4.4.0
urllib3 @ file:///croot/urllib3_1666298941550/work
Werkzeug==2.2.2
wrapt @ file:///tmp/abs_c335821b-6e43-4504-9816-b1a52d3d3e1eel6uae8l/croots/recipe/wrapt_1657814400492/work
yarl @ file:///opt/conda/conda-bld/yarl_1661437085904/work
zipp==3.10.0
